requirements.txt
Flask
opencv-python
mediapipe
numpy
mysql-connector-python
pymysql
boto3
